metrics: Use influxdata/t-digest for streaming quantiles
